+++ coloraide/color/__init__.py
@@ -26,8 +26,8 @@ from ..spaces.prophoto_rgb import ProPhotoRGB
 from ..spaces.rec2020 import Rec2020
 from ..spaces.xyz import XYZ
 from ..spaces.xyz_d50 import XYZD50
-from ..spaces.oklab import Oklab
-from ..spaces.oklch import Oklch
+from ..spaces.oklab.css import Oklab
+from ..spaces.oklch.css import Oklch
 from ..spaces.jzazbz import Jzazbz
 from ..spaces.jzczhz import JzCzhz
 from ..spaces.ictcp import ICtCp